requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# AI/ML
openai==1.3.3
httpx==0.27.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
feedparser==6.0.11
requests==2.31.0

# Environment & Utilities
python-dotenv==1.0.0
gunicorn==20.1.0

# NLP/Text Processing
nltk==3.8.1
